proximity2/backend/requirements.txt
# Backend Python dependencies for Proximity 2.0

# Django Core
Django==5.0.1
django-ninja==1.1.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0

# Async & Background Tasks
celery==5.3.6
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Proxmox Integration
proxmoxer==2.0.1
requests==2.31.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2

# Monitoring & Logging
sentry-sdk==1.39.2

# Development & Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.1.0

# Code Quality
black==23.12.1
isort==5.13.2
flake8==7.0.0
